[fix] Do NOT release buffer if Flatbuffers pointer is in use
According to the official flatbuffers tutorial, the `GetMonster`
is a convenience function that calls `GetRoot<Monster>`, which
just return the pointer to the root object inside the buffer. Therefore we should NOT release the buffer if the pointer is still
in use. (maybe applying "--gen-object-api" will introduce the copy operation.)